Split python side keyword modules to browser_control and browser_state. Rename ts side browserstate to browser_state
