+++ Browser/__init__.py
@@ -5,7 +5,7 @@ from robot.api import logger  # type: ignore
 from robot.libraries.BuiltIn import BuiltIn, EXECUTION_CONTEXTS  # type: ignore
 from robotlibcore import DynamicCore  # type: ignore
 
-from .keywords import Control, Getters, Input, Waiter
+from .keywords import Control, Getters, Input, State, Waiter
 from .playwright import Playwright
 from .version import VERSION
 
@@ -117,6 +117,7 @@ class Browser(DynamicCore):
         self.browser_control = Control(self)
         libraries = [
             self.browser_control,
+            State(self),
             Input(self),
             Getters(self),
             Waiter(self),
