+++ Browser/keywords/__init__.py
@@ -1,6 +1,7 @@
 from .browser_control import Control
+from .browser_state import State
 from .getters import Getters
 from .input import Input
 from .waiter import Waiter
 
-__all__ = ["Control", "Getters", "Input", "Waiter"]
+__all__ = ["Control", "State", "Getters", "Input", "Waiter"]
